% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_corr_tables.R
\name{make_corr_tables}
\alias{make_corr_tables}
\title{Make correlation tables}
\usage{
make_corr_tables(input_data, path_name = "~/", data_name = "data")
}
\arguments{
\item{input_data}{data.frame. The raw data.}

\item{path_name}{character. Where to put the correlation tables output csv files.}

\item{data_name}{character. What to name the correlation tables output csv files.}
}
\value{
corrs psych
}
\description{
Takes raw data and creates a correlation table and a correlation p value table. Outputs these tables to csv files.
}
\examples{
raw_data <- as.data.frame(readr::read_csv(
          system.file("extdata", "raw_data_for_correlation_calculation.csv", package = "nicertables")))
make_corr_tables(input_data = raw_data)
make_corr_tables(input_data = raw_data, path_name = "~/", data_name = "test_data")$r

}

