% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicertables-package.r
\docType{package}
\name{nicertables}
\alias{nicertables}
\alias{nicertables-package}
\title{nicertables: Make well-formatted analysis tables}
\description{
author: @lay_jennifer
}
\details{
This package contains functions that take raw data and badly-formatted analysis output and transforms it into results
tables that are ready to use for other purposes. There is currently one function that creates a correlation table from
raw data and outputs it to a csv file, and another function that takes logistical regression output from a totally
unformatted text document (Mplus output) and puts it into a data frame. These are convenience functions that are
probably most only useful for the package author...
}

